+++ 1.21.1/NeoForge/src/main/java/fuzs/puzzleslib/neoforge/impl/client/core/context/ResourcePackSourcesContextNeoForgeImpl.java
@@ -1,22 +1,31 @@
 package fuzs.puzzleslib.neoforge.impl.client.core.context;
 
-import com.google.common.base.Preconditions;
 import fuzs.puzzleslib.api.core.v1.context.PackRepositorySourcesContext;
+import fuzs.puzzleslib.neoforge.impl.core.context.DataPackSourcesContextNeoForgeImpl;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.PackType;
 import net.minecraft.server.packs.repository.RepositorySource;
+import net.neoforged.neoforge.event.AddPackFindersEvent;
 
 import java.util.Objects;
-import java.util.function.Consumer;
 
-public record ResourcePackSourcesContextNeoForgeImpl(
-        Consumer<RepositorySource> consumer) implements PackRepositorySourcesContext {
+public record ResourcePackSourcesContextNeoForgeImpl(AddPackFindersEvent event) implements PackRepositorySourcesContext {
 
     @Override
-    public void addRepositorySource(RepositorySource... repositorySources) {
-        Objects.requireNonNull(repositorySources, "repository sources is null");
-        Preconditions.checkState(repositorySources.length > 0, "repository sources is empty");
-        for (RepositorySource repositorySource : repositorySources) {
-            Objects.requireNonNull(repositorySource, "repository source is null");
-            this.consumer.accept(repositorySource);
-        }
+    public void registerRepositorySource(RepositorySource repositorySource) {
+        Objects.requireNonNull(repositorySource, "repository source is null");
+        this.event.addRepositorySource(repositorySource);
+    }
+
+    @Override
+    public void registerBuiltInPack(ResourceLocation resourceLocation, Component displayName, boolean isRequired) {
+        Objects.requireNonNull(resourceLocation, "resource location is null");
+        Objects.requireNonNull(displayName, "display name is null");
+        DataPackSourcesContextNeoForgeImpl.registerBuiltInPack(this.event,
+                resourceLocation,
+                displayName,
+                isRequired,
+                PackType.CLIENT_RESOURCES);
     }
 }