+++ 1.21.1/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.1.38-1.21.1] - 2025-08-11
+### Added
+- Add `PackRepositorySourcesContext::registerBuiltInPack`
+- Support built-in packs in `DataProviderHelper`
+
 ## [v21.1.37-1.21.1] - 2025-08-05
 ### Fixed
 - Fix `ShieldBlockCallback` being called even when not actively blocking on NeoForge